+++ glean-core/ios/GleanTests/GleanTests.swift
@@ -141,8 +141,88 @@ class GleanTests: XCTestCase {
         // Set up the expectation that will be fulfilled by the stub above
         expectation = expectation(description: "Baseline Ping Received")
 
+        // Set the last time the "metrics" ping was sent to now. This is required for us to not
+        // send a metrics pings the first time we initialize Glean and to keep it from interfering
+        // with these tests.
+        let now = Date()
+        Glean.shared.metricsPingScheduler.updateSentDate(now)
         // Restart Glean and don't clear the stores and then await the expectation
         Glean.shared.resetGlean(clearStores: false)
+        waitForExpectations(timeout: 5.0) { error in
+            XCTAssertNil(error, "Test timed out waiting for upload: \(error!)")
+        }
+    }
+
+    func testSendingDeletionPingIfDisabledOutsideOfRun() {
+        // Set up the test stub based on the default telemetry endpoint
+        let host = URL(string: Configuration.Constants.defaultTelemetryEndpoint)!.host!
+        stub(condition: isHost(host)) { data in
+            let path = (data as NSURLRequest).url!
+
+            let parts = path.absoluteString.split(separator: "/")
+
+            XCTAssertEqual("deletion-request", parts[4])
+
+            DispatchQueue.main.async {
+                // let the response get processed before we mark the expectation fulfilled
+                self.expectation?.fulfill()
+            }
+
+            return OHHTTPStubsResponse(
+                jsonObject: [],
+                statusCode: 200,
+                headers: ["Content-Type": "application/json"]
+            )
+        }
+
+        // Set up the expectation that will be fulfilled by the stub above
+        expectation = expectation(description: "Deletion Request Received")
+
+        // Reset Glean with uploadEnabled
+        Glean.shared.resetGlean(clearStores: true, uploadEnabled: true)
+
+        // Now reset Glean with uploadEnabled = false and not clearing the stores to
+        // trigger the deletion request ping.
+        Glean.shared.resetGlean(clearStores: false, uploadEnabled: false)
+        waitForExpectations(timeout: 5.0) { error in
+            XCTAssertNil(error, "Test timed out waiting for upload: \(error!)")
+        }
+    }
+
+    func testNotSendingDeletionRequestIfUnchangedOutsideOfRun() {
+        // Set up the test stub based on the default telemetry endpoint
+        let host = URL(string: Configuration.Constants.defaultTelemetryEndpoint)!.host!
+        stub(condition: isHost(host)) { _ in
+            XCTFail("Should not have recieved any ping")
+
+            DispatchQueue.main.async {
+                // let the response get processed before we mark the expectation fulfilled
+                self.expectation?.fulfill()
+            }
+
+            return OHHTTPStubsResponse(
+                jsonObject: [],
+                statusCode: 200,
+                headers: ["Content-Type": "application/json"]
+            )
+        }
+
+        // Set up the expectation that will NOT be fulfilled by the stub above.  If it is
+        // then it will trigger an assertion due to the `assertForOverFulfill` property.
+        expectation = expectation(description: "Deletion Request Received")
+
+        // So we can wait for expectations below, we will go ahead and fulfill the
+        // expectation.  We want to assert if the ping is triggered and over fulfills it
+        // from the stub above.
+        expectation?.fulfill()
+
+        // Reset Glean with uploadEnabled = false
+        Glean.shared.resetGlean(clearStores: true, uploadEnabled: false)
+
+        // Now reset Glean with uploadEnabled = false again without clearing the stores to
+        // make sure we don't trigger the deletion request ping.  If it does, then we will
+        // have overfulfilled the expectation which will trigger a test assertion.
+        Glean.shared.resetGlean(clearStores: false, uploadEnabled: false)
         waitForExpectations(timeout: 5.0) { error in
             XCTAssertNil(error, "Test timed out waiting for upload: \(error!)")
         }